requirements.txt
feedparser>=6.0
httpx>=0.27
pyyaml>=6.0
anthropic>=0.40
openai>=1.50
trafilatura>=1.12
fastapi>=0.115
uvicorn>=0.32
